“クエリ実行処理分離とディレクトリ名変更”
